Rework typings of explorer + aestetical changes
